restructure backend folders
