+++ lib/data/database/profile_database.dart
@@ -1,6 +1,6 @@
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
-import 'package:medication_tracker/model/user_profile_model.dart';
+import 'package:medication_tracker/data/model/user_profile_model.dart';
 
 class ProfileDatabaseHelper {
   static const _databaseName = "UserProfileDatabase.db";
