+++ lib/data/model/fda_drug_model.dart
