+++ lib/services/image/image_ui_handler.dart
@@ -4,10 +4,10 @@
 // image_service.dart
 import 'package:flutter/material.dart';
 
-import 'package:medication_tracker/camera_services/image_permission_handler.dart';
-import 'package:medication_tracker/model/medication_model.dart';
-import 'package:medication_tracker/providers/medication_provider.dart';
-import 'package:medication_tracker/providers/profile_provider.dart';
+import 'package:medication_tracker/services/image/image_permission_handler.dart';
+import 'package:medication_tracker/data/model/medication_model.dart';
+import 'package:medication_tracker/data/providers/medication_provider.dart';
+import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/utils/dialog_util.dart';
 import 'package:provider/provider.dart';
