+++ lib/ui/edit_medication_view.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:medication_tracker/camera_services/image_ui_handler.dart';
-import 'package:medication_tracker/model/medication_model.dart';
-import 'package:medication_tracker/providers/medication_provider.dart';
+import 'package:medication_tracker/services/image/image_ui_handler.dart';
+import 'package:medication_tracker/data/model/medication_model.dart';
+import 'package:medication_tracker/data/providers/medication_provider.dart';
 import 'package:medication_tracker/widgets/black_button.dart';
 import 'package:medication_tracker/widgets/header.dart';
 import 'package:medication_tracker/widgets/photo_upload_row.dart';
