+++ lib/ui/edit_profile_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:intl/intl.dart';
 import 'package:mask_text_input_formatter/mask_text_input_formatter.dart';
-import 'package:medication_tracker/model/user_profile_model.dart';
-import 'package:medication_tracker/providers/profile_provider.dart';
+import 'package:medication_tracker/data/model/user_profile_model.dart';
+import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/widgets/black_button.dart';
 import 'package:medication_tracker/widgets/header.dart';
 import 'package:medication_tracker/widgets/privacy_policy_button.dart';
