+++ lib/ui/select_profile_view.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:medication_tracker/providers/profile_provider.dart';
+import 'package:medication_tracker/data/providers/profile_provider.dart';
 import 'package:medication_tracker/ui/create_profile_view.dart';
 import 'package:medication_tracker/ui/home_view.dart';
 import 'package:medication_tracker/widgets/black_button.dart';
