+++ lib/widgets/med_tile.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:flutter/material.dart';
-import 'package:medication_tracker/model/medication_model.dart';
-import 'package:medication_tracker/providers/medication_provider.dart';
+import 'package:medication_tracker/data/model/medication_model.dart';
+import 'package:medication_tracker/data/providers/medication_provider.dart';
 import 'package:medication_tracker/ui/edit_medication_view.dart';
 import 'package:provider/provider.dart';
